Rename and move files for improved project structure
