+++ frontend/scripts/pages/home.page.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { Provider } from 'react-redux';
 import Home from 'react-components/home/home.container';
-import Nav from 'react-components/shared/nav/nav.container';
+import TopNavBar from 'react-components/shared/nav/top-nav-bar.container';
 import Footer from 'react-components/shared/footer.component';
 
 import 'styles/homepage.scss';
@@ -20,7 +20,7 @@ export const mount = (root, store) => {
 
   render(
     <Provider store={store}>
-      <Nav />
+      <TopNavBar />
     </Provider>,
     document.getElementById('nav')
   );
