+++ frontend/scripts/pages/profiles.page.jsx
@@ -7,7 +7,7 @@ import 'styles/profiles.scss';
 import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { Provider } from 'react-redux';
-import Nav from 'react-components/shared/nav/nav.container';
+import TopNavBar from 'react-components/shared/nav/top-nav-bar.container';
 import Footer from 'react-components/shared/footer.component';
 
 import values from 'lodash/values';
@@ -83,7 +83,7 @@ export const mount = (root, store) => {
 
   render(
     <Provider store={store}>
-      <Nav />
+      <TopNavBar />
     </Provider>,
     document.getElementById('nav')
   );
