+++ frontend/scripts/pages/tool.page.jsx
@@ -9,16 +9,16 @@ import { render, unmountComponentAtNode } from 'react-dom';
 import { Provider } from 'react-redux';
 import FlowContentContainer from 'containers/tool/tool-content.container';
 import SankeyContainer from 'containers/tool/sankey.container';
-import ColumnsSelectorContainer from 'containers/tool/columns-selector-react.container';
+import ColumnsSelectorContainer from 'react-components/tool/columns-selector.container';
 import MapContextContainer from 'containers/tool/map-context.container';
 import MapLegendContainer from 'containers/tool/map-legend.container';
 import MapBasemapsContainer from 'containers/tool/map-basemaps.container';
 import MapContainer from 'containers/tool/map.container';
-import NavReactContainer from 'containers/tool/nav-tool-react.container';
+import NavReactContainer from 'containers/../react-components/tool/nav-tool.container';
 import NavContainer from 'containers/tool/nav/nav-tool-navigation.container';
 import TitlebarContainer from 'containers/tool/titlebar.container';
 import NodesTitlesContainer from 'containers/tool/nodesTitles.container';
-import SearchContainer from 'containers/shared/search-react.container';
+import SearchContainer from 'react-components/shared/search/search.container';
 import ModalContainer from 'containers/tool/story-modal.container';
 import TooltipContainer from 'containers/shared/help-tooltip.container';
