+++ frontend/scripts/react-components/shared/search/node-title.component.jsx
