+++ frontend/scripts/react-components/shared/search/search-result.component.jsx
