+++ frontend/scripts/react-components/static-content/nav-sidebar.component.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import throttle from 'lodash/throttle';
 import { Transition } from 'react-transition-group';
-import NavLinksList from 'react-components/shared/nav-links-list.component';
+import NavLinksList from 'react-components/shared/nav/nav-links-list.component';
 
 class NavSidebar extends React.PureComponent {
   constructor(props) {
